Fix: Continuing format error in README... again...
Ok, this time I found an online .rst sandbox to test in. It looks
to me like the continued line of a bullet point must be in line
with the first character of the line above it. So, instead of
hitting tab, I have to enter 3 spaces. Why? Beats me.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>